wecom-bridge/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pyyaml==6.0.2
requests==2.32.3
python-multipart==0.0.20